Loan Application
This dataset includes details of applicants who have applied for loan. The dataset includes details like credit history, loan amount, their income, dependents etc.

Independent Variables:

Loan_ID - This refer to the unique identifier of the applicant's affirmed purchases
Gender - This refers to either of the two main categories (male and female) into which applicants are divided on the basis of their reproductive functions
Married - This refers to applicant being in a state of matrimony
Dependents - This refres to persons who depends on the applicants for survival
Education - This refers to number of years in which applicant received systematic instruction, especially at a school or university
Self_Employed - This refers to applicant working for oneself as a freelancer or the owner of a business rather than for an employer
ApplicantIncome - This refers to disposable income available for the applicant's use under State law.
CoapplicantIncome - This refers to disposable income available for the people that participate in the loan application process alongside the main applicant use under State law.
Loan_Amount - This refers to the amount of money an applicant owe at any given time.
Loan_Amount_Term - This refers to the duaration in which the loan is availed to the applicant
Credit History - This refers to a record of applicant's ability to repay debts and demonstrated responsibility in repaying them.
Property_Area - This refers to the total area within the boundaries of the property as set out in Schedule.
Loan_Status - This refres to whether applicant is eligible to be availed the Loan requested.